switch: all fixed

+++ engine/switch/timer.c
@@ -8,7 +8,6 @@
 
 #include <SDL.h>
 #include "timer.h"
-#include "types.h"
 
 #define GETTIME_FREQ (1000)
 
@@ -16,51 +15,42 @@ static unsigned lastinterval = 0;
 static u64 startcounter;
 unsigned newticks = 0;
 
-void borTimerInit()
-{
-	startcounter = SDL_GetPerformanceCounter();
+void borTimerInit() {
+    startcounter = SDL_GetPerformanceCounter();
 }
 
-void borTimerExit(){}
+void borTimerExit() {}
 
-unsigned timer_getinterval(unsigned freq)
-{
-	unsigned tickspassed,ebx,blocksize,now;
-	now=SDL_GetTicks()-newticks;
-	ebx=now-lastinterval;
-	blocksize=GETTIME_FREQ/freq;
-	ebx+=GETTIME_FREQ%freq;
-	tickspassed=ebx/blocksize;
-	ebx-=ebx%blocksize;
-	lastinterval+=ebx;
-	return tickspassed;
+unsigned timer_getinterval(unsigned freq) {
+    unsigned tickspassed, ebx, blocksize, now;
+    now = SDL_GetTicks() - newticks;
+    ebx = now - lastinterval;
+    blocksize = GETTIME_FREQ / freq;
+    ebx += GETTIME_FREQ % freq;
+    tickspassed = ebx / blocksize;
+    ebx -= ebx % blocksize;
+    lastinterval += ebx;
+    return tickspassed;
 }
 
-unsigned timer_gettick()
-{
-	return SDL_GetTicks();
+unsigned timer_gettick() {
+    return SDL_GetTicks();
 }
 
-u64 timer_uticks()
-{
-	u64 freq = SDL_GetPerformanceFrequency();
-	u64 counter = SDL_GetPerformanceCounter();
-	return counter * (1000000.0 / freq);
+u64 timer_uticks() {
+    u64 freq = SDL_GetPerformanceFrequency();
+    u64 counter = SDL_GetPerformanceCounter();
+    return (u64) (counter * (1000000.0 / freq));
 }
 
-unsigned get_last_interval()
-{
-	return lastinterval;
+unsigned get_last_interval() {
+    return lastinterval;
 }
 
-void set_last_interval(unsigned value)
-{
-	lastinterval = value;
+void set_last_interval(unsigned value) {
+    lastinterval = value;
 }
 
-void set_ticks(unsigned value)
-{
+void set_ticks(unsigned value) {
     newticks = value;
 }
-
-
